move manim/container/container.py to manim/container.py
